Move the warning to its own module
