+++ packages/app/src/cli/models/app/loader.ts
@@ -33,7 +33,7 @@ import {loadLocalExtensionsSpecifications} from '../extensions/load-specificatio
 import {UIExtensionSchemaType} from '../extensions/specifications/ui_extension.js'
 import {patchAppHiddenConfigFile} from '../../services/app/patch-app-configuration-file.js'
 import {getOrCreateAppConfigHiddenPath} from '../../utilities/app/config/hidden-app-config.js'
-import {showMultipleCLIWarningIfNeeded} from '@shopify/cli-kit/node/cli'
+import {showMultipleCLIWarningIfNeeded} from '@shopify/cli-kit/node/multiple-installation-warning'
 import {fileExists, readFile, glob, findPathUp, fileExistsSync} from '@shopify/cli-kit/node/fs'
 import {zod} from '@shopify/cli-kit/node/schema'
 import {readAndParseDotEnv, DotEnvFile} from '@shopify/cli-kit/node/dot-env'
@@ -351,7 +351,9 @@ class AppLoader<TConfig extends AppConfiguration, TModuleSpec extends ExtensionS
 
     const hiddenConfig = await loadHiddenConfig(directory, configuration)
 
-    await showMultipleCLIWarningIfNeeded(directory, nodeDependencies)
+    if (!this.previousApp) {
+      await showMultipleCLIWarningIfNeeded(directory, nodeDependencies)
+    }
 
     const {webs, usedCustomLayout: usedCustomLayoutForWeb} = await this.loadWebs(
       directory,
