+++ packages/cli-kit/package.json
@@ -92,12 +92,7 @@
                 "../../private/node/conf-store.js",
                 "../../private/node/constants.js",
                 "url",
-                "./cli-launcher.js",
-                "./is-global.js",
-                "./ui.js",
-                "./version.js",
-                "../common/version.js",
-                "./environment.js"
+                "./cli-launcher.js"
               ]
             }
           ]
